odds/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
base-scraper==1.0.2

charset-normalizer==3.4.4
cloudscraper==1.2.71
cryptography==46.0.3
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.1
fake-useragent==2.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
outcome==1.3.0.post0
packaging==25.0

postgrest==2.24.0
propcache==0.4.1
psycopg2-binary==2.9.11

pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
PySocks==1.7.1
python-decouple==3.8
realtime==2.24.0
redis==7.0.1
requests==2.32.5
requests-toolbelt==1.0.0
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
typing-inspection==0.4.2

undetected-chromedriver==3.5.5
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.22.0

gunicorn==23.0.0
